+++ feature/fido2/src/main/kotlin/com/chimali/fido2/presentation/navigation/Fido2RegistrationNavGraph.kt
@@ -64,9 +64,11 @@ fun Fido2RegistrationNavGraph(
         composable(Fido2Destinations.REGISTRATION_ROUTE) {
             RegistrationPromptScreen(
                 onSuccess = { credentialId ->
+                    navController.popBackStack()
                     onRegistrationComplete(credentialId)
                 },
                 onCancel = {
+                    navController.popBackStack()
                     onRegistrationCancelled()
                 }
             )